added upload and dl file

+++ ExpressJS/controllers/admin.js
@@ -1,4 +1,5 @@
 const { validationResult } = require('express-validator/check');
+const fileHelper = require('../util/file');
 
 const Product = require('../models/product');
 
@@ -21,8 +22,7 @@ exports.postAddProduct = (req, res, next) => {
   const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
-  console.log(description);
-  console.log(title);
+
   if (!image) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
@@ -38,6 +38,9 @@ exports.postAddProduct = (req, res, next) => {
       validationErrors: [],
     });
   }
+
+  const imgUrl = image.path;
+
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
@@ -59,6 +62,7 @@ exports.postAddProduct = (req, res, next) => {
   const product = new Product({
     title: title,
     price: price,
+    imgUrl: imgUrl,
     description: description,
     userId: req.user,
   });
@@ -128,8 +132,9 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImgUrl = req.body.imgUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
+
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
@@ -140,7 +145,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imgUrl: updatedImgUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId,
@@ -156,7 +160,10 @@ exports.postEditProduct = (req, res, next) => {
       }
       product.title = updatedTitle;
       product.price = updatedPrice;
-      product.imgUrl = updatedImgUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imgUrl);
+        product.imgUrl = image.path;
+      }
       product.description = updatedDesc;
       return product.save().then((result) => {
         console.log('UPDATED PRODUCT !!');
@@ -189,7 +196,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
+      fileHelper.deleteFile(product.imgUrl);
+      Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
